+++ pkg/renderers/containerv1alpha3/types.go
@@ -51,6 +51,19 @@ type EphemeralVolume struct {
 	ManagedStore string `json:"managedStore"`
 }
 
+func asEphemeralVolume(volume map[string]interface{}) (*EphemeralVolume, error) {
+	data, err := json.Marshal(volume)
+	if err != nil {
+		return nil, err
+	}
+	var ephemeralVolume EphemeralVolume
+	err = json.Unmarshal(data, &ephemeralVolume)
+	if err != nil {
+		return nil, err
+	}
+	return &ephemeralVolume, nil
+}
+
 func (ct ContainerTrait) MarshalJSON() ([]byte, error) {
 	properties := map[string]interface{}{}
 	for k, v := range ct.AdditionalProperties {
